Enhance logic to apply defaults
Ran into a problem where I wanted to specifically null a field that was non-null in the defaults class. Checking for a null value doesn't handle this scenario. Moved to the recently improved getPopulatedFieldsAsMap method to see what the user specifically assigned value (or absence of value) on the object and add values where the user hasn't specified anything.